+++ src/libmicroemu/machine.cpp
@@ -120,7 +120,7 @@ StatusCode Machine::Reset() noexcept {
   return StatusCode::kSuccess;
 }
 
-ExecResult Machine::Exec(i32 instr_limit, FPreExecStepCallback cb_pre_exec,
+ExecResult Machine::Exec(i64 instr_limit, FPreExecStepCallback cb_pre_exec,
                          FPostExecStepCallback cb_post_exec) noexcept {
   auto emu = BuildEmulator();
   auto res = emu.Exec(instr_limit, cb_pre_exec, cb_post_exec);
